% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PscanRBackgrounds-package.R
\docType{package}
\name{PscanRBackgrounds-package}
\alias{PscanRBackgrounds-package}
\alias{PscanRBackgrounds}
\title{Precomputed promoter backgrounds for PscanR}
\description{
PscanRBackgrounds provides ExperimentHub metadata and documentation for the
precomputed promoter-background score distributions used by the PscanR
motif enrichment package. Each background is a separate ExperimentHub
resource; see \link{PscanRBackgrounds-resources} for their content and
naming, and the package vignette for retrieval.
}
\examples{
path <- system.file("extdata", "metadata.csv", package = "PscanRBackgrounds")
metadata <- utils::read.csv(path)
nrow(metadata)
table(metadata$Species)
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/Federico77z/PscanRBackgrounds}
  \item \doi{10.5281/zenodo.21821764}
  \item Report bugs at \url{https://github.com/Federico77z/PscanRBackgrounds/issues}
}

}
\author{
\strong{Maintainer}: Federico Zambelli \email{federico.zambelli@unimi.it} (\href{https://orcid.org/0000-0003-3487-4331}{ORCID})

Authors:
\itemize{
  \item Federico Zambelli \email{federico.zambelli@unimi.it} (\href{https://orcid.org/0000-0003-3487-4331}{ORCID})
  \item Diana Betelli (\href{https://orcid.org/0009-0007-6207-7715}{ORCID})
  \item Giulio Pavesi (\href{https://orcid.org/0000-0001-5705-6249}{ORCID})
}

}
\keyword{internal}
