% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PscanRBackgrounds-package.R
\name{PscanRBackgrounds-package}
\alias{PscanRBackgrounds-package}
\title{Precomputed promoter backgrounds for PscanR}
\description{
PscanRBackgrounds provides ExperimentHub metadata and documentation for
immutable promoter-background score distributions used by PscanR.
}
\examples{
path <- system.file("extdata", "metadata.csv", package = "PscanRBackgrounds")
metadata <- utils::read.csv(path)
metadata[c("Title", "SourceVersion", "DispatchClass")]
}
\keyword{internal}
